+++ modules/@apostrophecms/piece-type/index.js
@@ -216,7 +216,6 @@ module.exports = {
       if (self.apos.launder.boolean(req.query['render-areas']) === true) {
         await self.apos.area.renderDocsAreas(req, result.results);
       }
-      self.apos.attachment.all(result.results, { annotate: true });
       if (query.get('choicesResults')) {
         result.choices = query.get('choicesResults');
       }
@@ -898,7 +897,7 @@ module.exports = {
         return piece;
       },
       getRestQuery(req) {
-        const query = self.find(req);
+        const query = self.find(req).attachments(true);
         query.applyBuildersSafely(req.query);
         if (!self.apos.permission.can(req, 'view-draft')) {
           if (!self.options.publicApiProjection) {